only authorized user can delete comment

+++ src/main/java/de/dhbw/mosbach/inf16a/kochbuch/commentservice/CommentController.java
@@ -5,7 +5,10 @@ import de.dhbw.mosbach.inf16a.kochbuch.authentication.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import de.dhbw.mosbach.inf16a.kochbuch.rezeptservice.*;
+import de.dhbw.mosbach.inf16a.kochbuch.authentication.*;
+import org.springframework.security.core.Authentication;
 
+import java.security.Principal;
 import java.util.List;
 
 /**
@@ -29,6 +32,11 @@ public class CommentController
 	@Autowired
 	private UserRepository userRepository;
 
+	@Autowired
+	private UserController userController;
+
+	@Autowired
+	private TokenManager tokenManager;
 
 	@GetMapping(value = "/comments")
 	public List<Comment> comments()
@@ -43,14 +51,22 @@ public class CommentController
 	}
 
 	@CrossOrigin
-	@DeleteMapping(value = "/comment/delete/{commentID}")
-	public void deleteCommeent(@PathVariable(value = "commentID") long commentID)
+	@DeleteMapping(value = "/comment/{commentID}")
+	public void deleteComment(@PathVariable(value = "commentID") long commentID, @RequestHeader Principal p)
 	// public Comment addComment(@RequestBody Comment theNewComment)
 	{
+		User user = userController.getUser(p);
+		Comment comment = commentRepository.findOne(commentID);
+
 		// Recipe re = recipeRepository.findOne(request.getRecipeId());
 		// User usr = userRepository.findOne(request.getUserId());
 		// Comment c = new Comment(request.getText(), request.getCreationDate(), usr, re);
-		commentRepository.delete(commentID);
+		if(user.getUserID() == comment.getUser().getUserID()){
+			commentRepository.delete(commentID);
+		}
+		else{
+			System.out.println("User: "+user.getUserID()+"not authorized to delete comment: "+commentID);
+		}
 		// return this.commentRepository.save(new Comment(request.getText(), request.getCreationDate(), usr, re));
 		// return commentRepository.findAllByOrderByCreationDateDesc();
 		// return c;
